requirements.txt
action-msgs==1.2.2
action-tutorials-interfaces==0.20.7
action-tutorials-py==0.20.7
actionlib-msgs==4.9.0
addict==2.4.0
ament-cmake-test==1.3.12
ament-copyright==0.12.14
ament-cppcheck==0.12.14
ament-cpplint==0.12.14
ament-flake8==0.12.14
ament-index-python==1.4.0
ament-lint==0.12.14
ament-lint-cmake==0.12.14
ament-package==0.14.1
ament-pep257==0.12.14
ament-uncrustify==0.12.14
ament-xmllint==0.12.14
anaconda==0.0.1.1
angles==1.15.0
appdirs==1.4.4
apturl==0.5.2
argcomplete==1.8.1
asttokens==3.0.1
attrs==25.4.0
Babel==2.8.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
blinker==1.9.0
Brlapi==0.8.3
Brotli==1.0.9
builtin-interfaces==1.2.2
catkin-pkg-modules==1.1.0
certifi==2020.6.20
chardet==4.0.0
chrome-gnome-shell==0.0.0
click==8.3.1
colcon-argcomplete==0.3.3
colcon-bash==0.5.0
colcon-cd==0.2.1
colcon-cmake==0.2.29
colcon-common-extensions==0.3.0
colcon-core==0.20.1
colcon-defaults==0.2.9
colcon-devtools==0.3.0
colcon-installed-package-information==0.2.1
colcon-library-path==0.2.1
colcon-metadata==0.2.5
colcon-notification==0.3.0
colcon-output==0.2.13
colcon-override-check==0.0.1
colcon-package-information==0.4.0
colcon-package-selection==0.2.10
colcon-parallel-executor==0.3.0
colcon-pkg-config==0.1.0
colcon-powershell==0.4.0
colcon-python-setup-py==0.2.9
colcon-recursive-crawl==0.2.3
colcon-ros==0.5.0
colcon-test-result==0.3.8
colcon-zsh==0.5.0
colorama==0.4.4
comm==0.2.3
command-not-found==0.3
composition-interfaces==1.2.2
ConfigArgParse==1.7.1
cov-core==1.15.0
coverage==6.2
cryptography==3.4.8
cupshelpers==1.0
cv-bridge==3.2.1
cycler==0.11.0
dash==3.3.0
dbus-python==1.2.18
debugpy==1.8.19
decorator==4.4.2
defer==1.0.6
demo-nodes-py==0.20.7
diagnostic-msgs==4.9.0
distlib==0.3.4
distro==1.7.0
distro-info==1.1+ubuntu0.2
docutils==0.17.1
domain-coordinator==0.10.0
duplicity==0.8.21
empy==3.3.4
example-interfaces==0.9.3
examples-rclpy-executors==0.15.5
examples-rclpy-minimal-action-client==0.15.5
examples-rclpy-minimal-action-server==0.15.5
examples-rclpy-minimal-client==0.15.5
examples-rclpy-minimal-publisher==0.15.5
examples-rclpy-minimal-service==0.15.5
examples-rclpy-minimal-subscriber==0.15.5
exceptiongroup==1.3.1
executing==2.2.1
fasteners==0.14.1
fastjsonschema==2.21.2
flake8==4.0.1
Flask==3.1.2
fonttools==4.29.1
fs==2.4.12
future==0.18.2
gast==0.5.2
geometry-msgs==4.9.0
html5lib==1.1
httplib2==0.20.2
idna==3.3
image-geometry==3.2.1
importlib-metadata==4.6.4
iniconfig==1.1.1
interactive-markers==2.3.2
ipykernel==7.1.0
ipython==8.37.0
ipywidgets==8.1.8
itsdangerous==2.2.0
jedi==0.19.2
jeepney==0.7.1
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.7.0
jupyter_core==5.9.1
jupyterlab_widgets==3.0.16
keyring==23.5.0
kiwisolver==1.3.2
language-selector==0.1
lark==1.1.1
laser-geometry==2.4.1
laspy==2.6.1
launch==1.0.13
launch-ros==0.19.13
launch-testing==1.0.13
launch-testing-ros==0.19.13
launch-xml==1.0.13
launch-yaml==1.0.13
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lifecycle-msgs==1.2.2
lockfile==0.12.2
logging-demo==0.20.7
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.1.3
map-msgs==2.1.0
Markdown==3.3.6
MarkupSafe==3.0.3
matplotlib==3.5.1
matplotlib-inline==0.2.1
mccabe==0.6.1
message-filters==4.3.11
monotonic==1.6
more-itertools==8.10.0
mpi4py==3.1.3
mpmath==0.0.0
narwhals==2.14.0
nav-msgs==4.9.0
nbformat==5.10.4
nest-asyncio==1.6.0
netifaces==0.11.0
nose2==0.9.2
notify2==0.3
numpy==1.25.2
numpy-stl==3.2.0
oauthlib==3.2.0
olefile==0.46
open3d==0.19.0
osrf-pycommon==2.1.6
packaging==25.0
pandas==2.3.3
paramiko==2.9.3
parso==0.8.5
pcl-msgs==1.0.0
pendulum-msgs==0.20.7
pexpect==4.8.0
pillow==12.0.0
platformdirs==4.5.1
plotly==6.5.0
pluggy==0.13.0
ply==3.11
plyfile==1.1.3
prompt_toolkit==3.0.52
protobuf==3.12.4
psutil==5.9.0
ptyprocess==0.7.0
pure_eval==0.2.3
py==1.10.0
pybind11==2.9.1
pycairo==1.20.1
pycodestyle==2.8.0
pycollada==0.9.2
pycups==2.0.1
pydocstyle==6.1.1
pydot==1.4.2
pyflakes==2.4.0
Pygments==2.11.2
PyGObject==3.42.1
pygraphviz==1.7
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-sip==12.9.1
pyquaternion==0.9.9
pyRFC3339==1.1
pytest==6.2.5
pytest-cov==3.0.0
python-apt==2.4.0+ubuntu4.1
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-qt-binding==1.1.3
python-utils==3.9.1
pythran==0.10.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==27.1.0
qt-dotgraph==2.2.4
qt-gui==2.2.4
qt-gui-cpp==2.2.4
qt-gui-py-common==2.2.4
quality-of-service-demo-py==0.20.7
rcl-interfaces==1.2.2
rclpy==3.3.19
rcutils==5.1.8
referencing==0.37.0
reportlab==3.6.8
requests==2.25.1
resource-retriever==3.1.3
retrying==1.4.2
rmw-dds-common==1.6.0
roman==3.3
ros2action==0.18.15
ros2bag==0.15.15
ros2cli==0.18.15
ros2component==0.18.15
ros2doctor==0.18.15
ros2interface==0.18.15
ros2launch==0.19.13
ros2lifecycle==0.18.15
ros2multicast==0.18.15
ros2node==0.18.15
ros2param==0.18.15
ros2pkg==0.18.15
ros2plugin==5.1.3
ros2run==0.18.15
ros2service==0.18.15
ros2topic==0.18.15
rosbag2-interfaces==0.15.15
rosbag2-py==0.15.15
rosdistro-modules==1.0.1
rosgraph-msgs==1.2.2
rosidl-adapter==3.1.8
rosidl-cli==3.1.8
rosidl-cmake==3.1.8
rosidl-generator-c==3.1.8
rosidl-generator-cpp==3.1.8
rosidl-generator-py==0.14.6
rosidl-parser==3.1.8
rosidl-runtime-py==0.9.3
rosidl-typesupport-c==2.0.2
rosidl-typesupport-cpp==2.0.2
rosidl-typesupport-fastrtps-c==2.2.4
rosidl-typesupport-fastrtps-cpp==2.2.4
rosidl-typesupport-introspection-c==3.1.8
rosidl-typesupport-introspection-cpp==3.1.8
rospkg-modules==1.6.1
rpds-py==0.30.0
rpyutils==0.2.2
rqt-action==2.0.1
rqt-bag==1.1.5
rqt-bag-plugins==1.1.5
rqt-console==2.0.3
rqt-graph==1.3.1
rqt-gui==1.1.9
rqt-gui-py==1.1.9
rqt-msg==1.2.0
rqt-plot==1.1.5
rqt-publisher==1.5.0
rqt-py-common==1.1.9
rqt-py-console==1.0.2
rqt-reconfigure==1.1.2
rqt-service-caller==1.0.5
rqt-shell==1.0.2
rqt-srv==1.0.3
rqt-topic==1.5.0
scikit-learn==1.7.2
scipy==1.8.0
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
sensor-msgs==4.9.0
sensor-msgs-py==4.9.0
shape-msgs==4.9.0
six==1.16.0
snowballstemmer==2.2.0
soupsieve==2.3.1
sros2==0.10.7
stack-data==0.6.3
statistics-msgs==1.2.2
std-msgs==4.9.0
std-srvs==4.9.0
stereo-msgs==4.9.0
sympy==1.9
systemd-python==234
teleop-twist-keyboard==2.4.1
tf2-geometry-msgs==0.25.18
tf2-kdl==0.25.18
tf2-msgs==0.25.18
tf2-py==0.25.18
tf2-ros-py==0.25.18
tf2-sensor-msgs==0.25.18
tf2-tools==0.25.18
threadpoolctl==3.6.0
toml==0.10.2
topic-monitor==0.20.7
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
trajectory-msgs==4.9.0
trimesh==4.10.1
turtlesim==1.4.3
typeguard==2.2.2
typing_extensions==4.15.0
tzdata==2025.3
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
unique-identifier-msgs==2.2.1
urllib3==1.26.5
usb-creator==0.3.7
visualization-msgs==4.9.0
wadllib==1.3.6
wcwidth==0.2.14
webencodings==0.5.1
Werkzeug==3.1.4
widgetsnbextension==4.0.15
xatlas==0.0.11
xdg==5
xkit==0.0.0
zipp==1.0.0
